Increased display-panes-timeout to forever.
When you press `^bq`, the display-panes command, a number appears in
every pane in the window. While the numbers are displayed, you can press
a number (0-9) and switch to the pane. This feature is really handy and
can be used to avoid a lot of `^b<arrow-key>`'s. However, tmux only
shows the numbers for a short amount of time by default. I increased the
timeout to some very large number that I never anticipate hitting. Also,
if you want to cancel a display-panes command, you can simple press
escape.